Initial webpack setup
This installs the `webpacker` and moves
around all our existing JS to work with the
new asset pipeline.

An exception to this is `rails_stack_view`, which
is included in the traditional way.

The `.browerslistrc` has been configured
to target IE11 which is our oldest supported
browser.

This also updates GitHub actions to run `rake assests:precompile`
which will run `yarn install` and `webpack` to pull in the
dependencies.