LAG-3993 Google Analytics JS Error
This removes a script from out application.js pack that uses an
older method of tracking events in GA. Right now it is
causing an error because it refers to JS object that our
GA script no longer provides.

If we still want to track these events this will
need to rewritten to use this:

https://developers.google.com/analytics/devguides/collection/gtagjs/sending-data

Some of these events refer to UI that is no longer present in
the B7 update.