LAG-4068 LAG-4069 Advanced search/quick search behavior
This simplifies the creation of the advanced search link
based on changes from LAG-4050. This avoids a
search term always being in the `Any Field` field
in advanced search even if you selected title from
the dropdown.

This overrides the default blacklight label behavior
so that we show the Any Field label on results pages
for the quick search/default search.