LAG-3997 Catalyst stored searches (#90)
This adds a nightly task using `whenever` to
clear out saved searches.The cron tab
entries are installed on the server when the
application is deployed with `capistrano`.

It looks like tasks and the capistrano/whenever
setup existed before, but weren't migrated when
the move to `ansible` happened.

I removed the existing tasks for cleaning out searches
and users. As we accumulate users we will need to
reinstate a similar task, but we will need to be careful
because we won't be able to use `TRUNCATE` like we are
with searches.