Rails 6.1 Upgrade / Ruby 3
This commit upgrades the Rails version to 6.1.3.2 and
sets the ruby version to 3 in the Gemfile.

The Gemfile has also been cleaned up.

There was a failing test for a relative time zone helper
because it was using a deprecated method. The helper has
been changed to work around that.