LAGSCRUM-196 Unrelated "Online Access" link appearing in map records (#166)
This adds a helper that checks the format of the Solr document being
displayed. If it has the 'Map/Globe' format it won't display FindIt URLs
or load the FindIt JS on the item page.

FindIt was sending maps to SFX as books. We don't have map coverage in SFX
so this ensures we don't look anyway.