LAGSCRUM-326 Explore HathiTrust API
This adds an endpoint that we can use to query
the HathiTrust API. Right now it is only using
the OCLC number the lookup records.

The endpoint can't be accessed directly. It's
intended to be used via JS eventually. To use
it you need to have a browser session. You
don't need to be logged in, but a session is
required.