LAGSCRUM-548 Implement workflow to overlay subject headings in Catalyst
This adds a class called `SubjectOverrider` that's used
for overriding subjects based on the translation
map present in `config/translation_maps/subject_overrides.yml`.

A yaml file that has the same name should also be edited in the `catalyst_traject`
repo when making changes to the map.

Catalyst doesn't display values directly from Solr. It displays
content from a binary MARC record stored in Solr. This class
is used to translate the subjects that are displayed from that
record.